ollama-langfuse/requirements.txt
langfuse>=2.53.3
openai>=1.0.0
python-dotenv
requests